refactor(shared): extract env-sanitize.ts — consolidate child-env/runtime duplication
Extract 5 pure env sanitization functions (resolveEnvBudgetBytes, parsePassthrough,
measureEnv, shedIfOverBudget, stripInternalCoralKeys) into shared/env-sanitize.ts.

Both child-env.ts (claude-appserver consumer) and runtime.ts (backend daemon)
now delegate to the shared module. Removes ~70 lines of duplicated ARG_MAX
detection and env shedding logic from runtime.ts.
